requirements.txt
fastapi==0.116.1
pydantic==2.11.7
python-multipart==0.0.20
qrcode==8.2
pillow==11.3.0
python-docx==1.2.0
pypdf==5.9.0
reportlab==4.4.3
